+++ expense-app/src/components/Tabs/tab.scss
@@ -96,8 +96,10 @@ button {
 .chart-container {
   height: 424px;
   width: 428px;
+  margin: 0 auto;
 }
 .dougnut-container {
   height: 424px;
   width: 428px;
+  margin: 0 auto;
 }
