+++ expense-app/src/helpers/computeForDougnutChart.ts
@@ -1,0 +1,1 @@
+export const income = () => {}
