+++ expense-app/src/helpers/computeForPieChart.ts
@@ -22,3 +22,36 @@ export const getExpenses = (transactions: Transaction[]) => {
     });
     return expense;
 }
+
+export const PieChartData = (transactions: Transaction[]) => {
+    const income = getIncome(transactions).reduce((a , b) => a + b);
+    const expense = getExpenses(transactions).reduce((a , b) => a + b);
+    return {
+        labels: ['Income', 'Expenses'],
+        datasets: [
+          {
+            label: '# of Votes',
+            data: [income, expense],
+            backgroundColor: [
+              'rgba(255, 99, 132, 0.2)',
+              'rgba(54, 162, 235, 0.2)',
+              'rgba(255, 206, 86, 0.2)',
+              'rgba(75, 192, 192, 0.2)',
+              'rgba(153, 102, 255, 0.2)',
+              'rgba(255, 159, 64, 0.2)',
+            ],
+            borderColor: [
+              'rgba(255, 99, 132, 1)',
+              'rgba(54, 162, 235, 1)',
+              'rgba(255, 206, 86, 1)',
+              'rgba(75, 192, 192, 1)',
+              'rgba(153, 102, 255, 1)',
+              'rgba(255, 159, 64, 1)',
+            ],
+            borderWidth: 1,
+          },
+        ]
+    }
+
+   
+  };
